Move Service Registry into lib and cut down intializer to aid testing
