+++ config/initializers/service_registry.rb
@@ -1,21 +1,3 @@
-module DefenceRequestServiceRota
-  class ServiceNotRegistered < ArgumentError; end
-
-  def self.service(name)
-    services.fetch(name).new { raise_missing_service_error(name) }
-  end
-
-  def self.register_service(name, service)
-    services[name] = service
-  end
-
-  def self.raise_missing_service_error(name)
-    raise ServiceNotRegistered, "Service #{name} is not registered"
-  end
-
-  def self.services
-    @services ||= {}
-  end
-end
+require Rails.root.join("lib/service_registry")
 
 DefenceRequestServiceRota.register_service(:auth_api, FakeDataApi)
