+++ lib/service_registry.rb
@@ -1,0 +1,19 @@
+module DefenceRequestServiceRota
+  class ServiceNotRegistered < ArgumentError; end
+
+  def self.service(name)
+    services.fetch(name).new { raise_missing_service_error(name) }
+  end
+
+  def self.register_service(name, service)
+    services[name] = service
+  end
+
+  def self.raise_missing_service_error(name)
+    raise ServiceNotRegistered, "Service #{name} is not registered"
+  end
+
+  def self.services
+    @services ||= {}
+  end
+end
